Restructure devtools project
