+++ noxfile.py
@@ -5,7 +5,7 @@
 # License: FuzzyLite License
 # Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
 
-"""fuzzylite-devtools: developer tools to work on fuzzylite
+"""fuzzylite-devtools: developer tools for fuzzylite
 
 by Juan Rada-Vilela, PhD.
 
@@ -20,258 +20,24 @@ Examples:
 
 from __future__ import annotations
 
-import dataclasses
-import datetime
-import json
-import os
-import platform
-import re
-import shutil
-import tempfile
-import textwrap
-from pathlib import Path
-from typing import Any, Literal
-
 import nox
-import tqdm
-from nox.command import CommandFailed
+import platform
+import shutil
+from fuzzylite_devtools import Configuration, Tools
+from pathlib import Path
 
 # use the virtual environment from which nox was called
 nox.options.force_venv_backend = "none"
-nox.options.sessions = ["version", "configure", "build", "test", "coverage"]
-
-
-Build = Literal["release", "debug", "relwithdebinfo", "minsizerel"]
-CxxStandard = Literal["98", "11", "14", "17", "20", "23", "26"]
-
-
-@dataclasses.dataclass
-class Configuration:
-    """Configuration class for fuzzylite."""
-
-    build: Build = "relwithdebinfo"
-    cxx_standard: CxxStandard = "11"
-    install_prefix: str = ".local"
-    strict: bool = False
-    tests: bool = True
-    coverage: bool = True
-    use_float: bool = False
-    jobs: int = max((os.cpu_count() or 1) // 3, 1)
-    container: str = "docker"
-    _posargs: list[str] = dataclasses.field(default_factory=list)
-
-    def posargs(self) -> str:
-        return " ".join(self._posargs)
-
-    def build_path(self) -> Path:
-        """Return the build path (eg, `build/release`)
-
-        Returns:
-            the build path
-        """
-        return Tools.base_build() / self.build
-
-    def convert_value(self, property: str, value: Any) -> Any:
-        """Convert the value based on the current contents of the configuration
-        Args:
-            property: name of configuration property
-            value: value of the configuration property
-
-        Returns:
-            the given value converted to the appropriate type
-
-        Raises:
-            ValueError, when the conversion fails or the type of value is not handled.
-        """
-        current_value = getattr(self, property)
-        if isinstance(current_value, str):
-            return value
-        if isinstance(current_value, bool):
-            boolean = {
-                "true": True,
-                "on": True,
-                "1": True,
-                "false": False,
-                "off": False,
-                "0": False,
-            }
-            return boolean[str(value).lower()]
-        if isinstance(current_value, int):
-            return int(value)
-        raise ValueError(f"unknown type of value: {value} ({current_value})")
-
-    def override(self, posargs: list[str]) -> Configuration:
-        """Override the configuration with the session arguments
-        The session arguments can be known key-value pairs, like `cxx_standard=20` or `--verbose`
-
-        Args:
-            posargs: the session arguments.
-
-        Returns:
-            self with updated properties based on the session arguments
-
-        Raises:
-            ValueError, when the conversion of known `key=value` pairs fails
-        """
-        for arg in posargs:
-            try:
-                key, value = arg.split("=")
-            except ValueError:
-                self._posargs.append(arg)
-                continue
-            if hasattr(self, key):
-                value = self.convert_value(key, value)
-                setattr(self, key, value)
-            else:
-                self._posargs.append(arg)
-        return self
-
-    def to_env(self) -> str:
-        export = [
-            f"{key}={value}"
-            for key, value in vars(self).items()
-            if not key.startswith("_")
-        ]
-        if self.posargs():
-            export.append(f"posargs={self.posargs()}")
-        return "\n".join(export)
-
-    def to_json(self) -> str:
-        export = {
-            key: value for key, value in vars(self).items() if not key.startswith("_")
-        }
-        return json.dumps(export)
-
-    @staticmethod
-    def from_json(json_text: str) -> Configuration:
-        return Configuration(**json.loads(json_text))
-
-    def save(self) -> None:
-        """Save the configuration file"""
-        Tools.configuration_file().write_text(self.to_json())
-
-    @staticmethod
-    def load() -> Configuration:
-        """Load the configuration file"""
-        return Configuration.from_json(Tools.configuration_file().read_text())
-
-    @staticmethod
-    def for_session(
-        session: nox.Session, load_env: bool = True, posargs: bool = True
-    ) -> Configuration:
-        """Load the default configuration,
-        overriding with the configuration file
-        and the session arguments.
-
-        Args:
-            session: task with optional arguments
-            load_env: whether to override default configuration with the `build/fuzzylite.env` environment file
-            posargs: whether to override the configuration with the session arguments
-
-        Returns:
-            a configuration with properties overriden from environment file and session
-        """
-        if load_env and Tools.configuration_file().exists():
-            configuration = Configuration.load()
-        else:
-            configuration = Configuration()
-
-        if posargs:
-            configuration.override(session.posargs)
-        session.log(
-            "\nConfiguration:\n" + textwrap.indent(configuration.to_env(), prefix="\t")
-        )
-        return configuration
-
-
-class Tools:
-    @staticmethod
-    def base_build() -> Path:
-        """Return the path to the base build folder (ie, `./build`)
-
-        Returns:
-            Path to `./build`
-        """
-        return Path("./build")
-
-    @staticmethod
-    def configuration_file() -> Path:
-        return Tools.base_build() / "fuzzylite.json"
-
-    @staticmethod
-    def poetry_directory() -> Path:
-        """Return the path to the poetry project
-
-        Returns:
-            Path to `./tools.dev`
-        """
-        return Path("./tools/dev")
-
-    @staticmethod
-    def create_temporal_directory(name: str) -> Path:
-        """Create a temporal directory in the form `com.fuzzylite.{name}.{yyyy/MM/dd}`
-        Args:
-            name: partial name of the temporal directory
-
-        Returns:
-            a temporal directory
-        """
-        now = re.sub(
-            r"\W", "", datetime.datetime.now().replace(microsecond=0).isoformat()
-        )
-        temporal_directory = tempfile.mkdtemp(prefix=f"com.fuzzylite.{name}.{now}.")
-        return Path(temporal_directory)
-
-    @staticmethod
-    def source_files() -> list[Path]:
-        """Find C++ header and source files in the project
-
-        Returns:
-            C++ header and source files
-        """
-        paths = ["fuzzylite/", "fl/", "src/", "test/"]
-        files = []
-        for path in paths:
-            files.extend(list(Path(path).rglob("**/*.h")))
-            files.extend(list(Path(path).rglob("**/*.cpp")))
-        files.sort()
-        return files
-
-    @staticmethod
-    def clang_format(mode: Literal["lint", "format"], session: nox.Session) -> None:
-        """Run the `clang-format` tool on each header and source file of the project.
-
-        Args:
-            mode: `lint` will check the files, and `format` will format the files in place
-            session: to log progress
-        """
-        clang_format_cmd = "clang-format --style=file:.clang-format"
-        parameters = {"format": "-i", "lint": "--dry-run --Werror"}
-        message = {"format": "Formatting", "lint": "Linting"}
-        files = Tools.source_files()
-        errors: list[Path] = []
-        session.log(f"{message[mode]} {len(files)} files...")
-        with tqdm.tqdm(files, desc=f"{mode}") as bar:
-            for file in bar:
-                bar.set_description(f"{mode}: {file.parent}".ljust(30))
-                cmd = " ".join([clang_format_cmd, parameters[mode], str(file)])
-                try:
-                    session.run(*cmd.split(), log=False)
-                except CommandFailed:
-                    errors.append(file)
-                except (KeyboardInterrupt, Exception):  # CTRL+C or else
-                    break
-            bar.set_description(None)
-        if errors:
-            error_messages = "\n".join(str(file) for file in errors)
-            session.error(f"The following files need formatting:\n{error_messages}")
+nox.options.sessions = ["initialize", "configure", "build", "test", "coverage"]
 
 
 ## Sessions
 @nox.session
-def version(session: nox.Session) -> None:
-    """Print the relevant libraries and their versions."""
+def initialize(session: nox.Session) -> None:
+    """Initialize developer tools"""
     session.log(platform.python_version())
+    session.run(*"poetry config virtualenvs.create false".split())
+    session.run(*f"poetry lock -C {Tools.poetry_directory()}".split())
     session.run(*f"poetry show -T -C {Tools.poetry_directory()}".split())
 
 
@@ -378,7 +144,7 @@ def docs(session: nox.Session) -> None:
 def clean(session: nox.Session) -> None:
     """Clean the project (default `all`). Args: `[all|last|coverage]` to remove `build/` folder, last build (eg, `build/release`),  and *.gcda coverage files."""
     c = Configuration.for_session(session)
-    args = set(c.posargs())
+    args = set(c.posargs().split())
     error_message = (
         f"expected one of {['all', 'last', 'coverage']}, but found: {c.posargs()}"
     )
@@ -399,7 +165,7 @@ def clean(session: nox.Session) -> None:
     if build_path.exists():
         tempdir = Tools.create_temporal_directory("build")
         shutil.move(build_path, tempdir)
-        session.log(f"moved {build_path} to {tempdir}")
+        session.log(f"moved '{build_path}' to '{tempdir}'")
     else:
         session.log(f"nothing to clean: '{build_path}' does not exist")
 
@@ -453,8 +219,6 @@ def lint_py(session: nox.Session) -> None:
     session.notify(lint_py_ruff.__name__)
     session.notify(lint_py_right.__name__)
     session.run(*f"poetry check -C {Tools.poetry_directory()}".split())
-
-
 
 
 @nox.session
