+++ tools/dev/fuzzylite_devtools/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Configuration", "Tools"]
+
+from .configuration import Configuration, Tools
